Небольшие правки.

+++ Shapes/src/com/git/stat1q/shapes/Circle.java
@@ -48,6 +48,6 @@ public class Circle implements Shape {
     public int hashCode() {
         final int prime = 37;
         int hash = 1;
-        return Double.hashCode(prime * hash + (int) radius);
+        return Double.hashCode(prime * hash + radius);
     }
 }
